requirements.txt
flask
flask_sqlalchemy
werkzeug
flask_jwt_extended
sqlalchemy
together
dotenv
flask_cors
azure-ai-inference